move PerfTestApp* out of smoketest; it's not a smoketest app
