--- test/smoke/testApps/PerfTestAppSpringBoot/src/main/java/com/microsoft/applicationinsights/testapps/perf/TestCaseRunnableFactory.java
@@ -1,19 +1,0 @@
-package com.microsoft.applicationinsights.testapps.perf;
-
-public abstract class TestCaseRunnableFactory {
-
-    private final String name;
-
-    public TestCaseRunnableFactory() {
-        this(null);
-    }
-    public TestCaseRunnableFactory(String name) {
-        this.name = name;
-    }
-
-    public TestCaseRunnable get() {
-        return new TestCaseRunnable(getRunnable(), name);
-    }
-
-    protected abstract Runnable getRunnable();
-}
