+++ test/webapps/PerfTestApp/src/main/resources/ApplicationInsights.xml
