+++ test/webapps/PerfTestAppSpringBoot/src/main/resources/application.properties
@@ -1,0 +1,9 @@
+spring.application.name=PerfTestAppSpringBoot
+
+azure.application-insights.instrumentation-key=00000000-0000-0000-0000-cba987654321
+azure.application-insights.logger.level=INFO
+
+spring.mvc.favicon.enabled=false
+
+spring.mvc.view.prefix=/WEB-INF/jsp/
+spring.mvc.view.suffix=.jsp
